Create src folder. Move stuff around.
