+++ test/test_KnittingConversions.py
@@ -1,6 +1,6 @@
 import sys
 sys.path.append('..')
-from KnittingConversions import *
+from conversions import *
 import unittest
 from collections import Counter
 from math import floor, ceil
